GameOver fixes
implemented a holdPopUp method that just changes the popUp without the fade animation. Used for when the game is lost and the final word is revealed.

Game now cuts to the Stats Page when the game is finished.

Removed some redundant System.out.println(); messages (now implemented in the gui so no use)